+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-*
+* Warning log when `HTTPClient` synchronous call is made on an already running event loop
 
 ### Changed